[backend.authentication] - удаление функционала связанного в only user [backend.user] - создание проекта, с переносом в него всего user only функционала
