+++ backend/authentication/schemas.py
@@ -3,25 +3,6 @@ import datetime
 from pydantic import BaseModel, EmailStr
 
 
-class BaseUserSchema(BaseModel):
-    username: str
-    email: EmailStr
-
-
-class UserSchemaGet(BaseUserSchema):
-    id: str
-
-
-class UserSchemaCreate(BaseUserSchema):
-    password: str
-
-
-class UpdateUserSchema(BaseModel):
-    username: str | None
-    email: EmailStr | None
-    password: str | None
-
-
 class UpdatePasswordSchema(BaseModel):
     password: str
