+++ backend/main.py
@@ -7,6 +7,7 @@ from backend.core.database.driver import init_database_client
 from backend.authentication.urls import app_authentication
 from backend.core.middleware import authentication_middleware
 from backend.tree.urls import app_tree
+from backend.user.urls import app_user
 
 origins = [
     "http://127.0.0.1:3000"
@@ -18,6 +19,7 @@ app = FastAPI(middleware=[
 ])
 
 app.mount('/authentication', app_authentication)
+app.mount('/user', app_user)
 app.mount('/tree', app_tree)
