+++ backend/user/dependence.py
@@ -1,0 +1,5 @@
+from backend.user.crud import UserCrud
+
+
+def user_crud():
+    return UserCrud()
