+++ backend/user/test/test_api.py
@@ -3,7 +3,7 @@ from unittest import IsolatedAsyncioTestCase
 
 from httpx import AsyncClient
 
-from backend.authentication.crud import UserCrud
+from backend.user.crud import UserCrud
 from backend.core.database.driver import init_database_client, get_database
 from backend.main import app
 
@@ -22,7 +22,7 @@ class UserCrudTestCase(IsolatedAsyncioTestCase):
     async def test_get_user(self):
         user = await self.crud.create({'username': 'pushkin', 'email': 'pushkin@mail.com', 'password': b'veverbi344'})
         async with AsyncClient(app=app, base_url="http://127.0.0.1") as ac:
-            response = await ac.get(f"/authentication/{user}")
+            response = await ac.get(f"/user/{user}")
         self.assertEqual(response.status_code, 200)
         user_data = json.loads(response.content)
         self.assertEqual(user_data['username'], 'pushkin')
@@ -30,5 +30,5 @@ class UserCrudTestCase(IsolatedAsyncioTestCase):
 
     async def test_get_not_exist_user(self):
         async with AsyncClient(app=app, base_url="http://127.0.0.1") as ac:
-            response = await ac.get(f"/authentication/6231c1098d937d9ce2da8f20")
+            response = await ac.get(f"/user/6231c1098d937d9ce2da8f20")
         self.assertEqual(response.status_code, 204)
